+++ charts/core/src/index.ts
@@ -1,5 +1,4 @@
 export { Chart, type ChartProps, ChartStates } from './Chart';
-export { ChartCard, type ChartCardProps, ChartCardStates } from './ChartCard';
 export {
   EventMarkerLine,
   type EventMarkerLineProps,
@@ -10,12 +9,6 @@ export { Grid, type GridProps } from './Grid';
 export { Header, type HeaderProps } from './Header';
 export { Line, type LineProps } from './Line';
 export { ThresholdLine, type ThresholdLineProps } from './ThresholdLine';
-export {
-  SortDirection,
-  SortKey,
-  SortOrder,
-  Tooltip,
-  type TooltipProps,
-} from './Tooltip';
+export { type SeriesInfo, Tooltip, type TooltipProps } from './Tooltip';
 export { XAxis, type XAxisProps, type XAxisType } from './XAxis';
 export { YAxis, type YAxisProps, type YAxisType } from './YAxis';