+++ tools/codemods/src/index.ts
@@ -6,6 +6,7 @@ import * as jscodeshift from 'jscodeshift/src/Runner';
 import path from 'path';
 
 import { checkGitStatus } from './utils/checkGitStatus';
+import { getAvailableCodemods } from './utils/getAvailableCodemods';
 import { LGPackage } from './types';
 
 export interface MigrateOptions {
@@ -14,6 +15,7 @@ export interface MigrateOptions {
   force?: boolean;
   ignore?: Array<string>;
   packages?: Array<LGPackage>;
+  list?: boolean;
 }
 
 export const migrator = async (
@@ -21,6 +23,38 @@ export const migrator = async (
   files?: string | Array<string>,
   options: MigrateOptions = {},
 ) => {
+  if (options.list || codemod === 'list') {
+    // If list option is true, (or the provided codemod name is "list")
+    // show available codemods and exit
+    const codemods = await getAvailableCodemods();
+
+    if (codemods.length === 0) {
+      console.log(chalk.yellow('No codemods found'));
+    } else {
+      console.log(chalk.greenBright('Available codemods:'));
+      codemods.forEach(cm => {
+        console.log(`- ${chalk.cyan(cm)}`);
+      });
+      console.log(
+        `\nFor more details visit: `,
+        chalk.gray(
+          '\nhttps://github.com/mongodb/leafygreen-ui/blob/main/tools/codemods/README.md#codemods-1',
+        ),
+      );
+    }
+
+    return;
+  }
+
+  // Check if a valid codemod name is provided
+  if (!codemod || codemod.trim() === '') {
+    console.error(chalk.red('Error: No codemod specified'));
+    console.log(
+      chalk.yellow('Use the --list option to see available codemods.'),
+    );
+    process.exit(1);
+  }
+
   let _files = files;
   // Gets the path of the codemod e.g: /Users/.../leafygreen-ui/tools/codemods/dist/codemod/[codemod]/transform.js
   const codemodFile = path.join(
@@ -33,7 +67,7 @@ export const migrator = async (
   try {
     if (!fse.existsSync(codemodFile)) {
       throw new Error(
-        `No codemod found for ${codemod}. The list of codemods can be found here: https://github.com/mongodb/leafygreen-ui/blob/main/tools/codemods/README.md#codemods-1`,
+        `No codemod found named "${codemod}". Use --list to see available codemods.`,
       );
     }