+++ tools/create/src/index.ts
@@ -1,25 +1,39 @@
-import { getLGConfig } from '@lg-tools/meta';
+/* eslint-disable no-console */
+import chalk from 'chalk';
 
+import { getComponentRootDirectory } from './utils/getComponentRootDirectory';
+import { getScope } from './utils/getScope';
 import { CreatePackageOptions } from './create.types';
 import { createComponent } from './createComponent';
 import { createSubComponent } from './createSubComponent';
 
 export function createPackage(name: string, options: CreatePackageOptions) {
-  const { scopes } = getLGConfig();
-
   const parent = options.parent;
 
-  // TODO: get scope & directory from parent
-  const scope = options.scope ?? Object.keys(scopes)[0];
-  const directory = options.directory ?? scopes[scope];
+  const scope = getScope(options.scope, options.parent);
+  const directory = getComponentRootDirectory({
+    scope,
+    parent,
+    directory: options.directory,
+  });
+
+  options.verbose && console.log(chalk.bold('Creating package:'), name);
+  options.verbose && console.log(chalk.bold('Scope:'), scope);
+  options.verbose && console.log(chalk.bold('Root Directory:'), directory);
+  options.verbose && console.log(chalk.bold('Parent component:'), parent);
 
   if (parent) {
-    createSubComponent({ name, parent });
+    createSubComponent({
+      ...options,
+      name,
+      parent,
+    });
   } else {
     createComponent({
-      name,
+      ...options,
       scope,
       directory,
+      name,
     });
   }
 }