+++ tools/create/src/utils/getNameVariants.ts
@@ -3,10 +3,12 @@ import kebabCase from 'lodash/kebabCase';
 import startCase from 'lodash/startCase';
 
 export function getNameVariants(name: string) {
+  const component = name.split('/')[1] ?? name;
+
   // Construct all required parameters
-  const packageNameKebab = kebabCase(name);
-  const packageNameTitle = startCase(name);
-  const packageNamePascal = camelCase(name).replace(/^\w/, c =>
+  const packageNameKebab = kebabCase(component);
+  const packageNameTitle = startCase(component);
+  const packageNamePascal = camelCase(component).replace(/^\w/, c =>
     c.toUpperCase(),
   );