+++ charts/colors/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @lg-charts/colors
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [30b13adec]
+  - @leafygreen-ui/lib@14.2.0
+  - @leafygreen-ui/leafygreen-provider@4.0.6
+  - @leafygreen-ui/palette@4.1.4
+
+## 0.2.2
+
+### Patch Changes
+
+- 206278b0f: Update charts color palette
+
 ## 0.2.1
 
 ### Patch Changes