+++ charts/legend/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @lg-charts/legend
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [30b13adec]
+- Updated dependencies [78a36d6bb]
+  - @leafygreen-ui/lib@14.2.0
+  - @leafygreen-ui/emotion@4.1.0
+  - @leafygreen-ui/checkbox@15.2.1
+  - @leafygreen-ui/leafygreen-provider@4.0.6
+  - @lg-charts/colors@0.2.3
+  - @leafygreen-ui/palette@4.1.4
+  - @leafygreen-ui/tokens@2.12.1
+  - @lg-charts/series-provider@0.2.1
+
+## 0.2.0
+
+### Minor Changes
+
+- 206278b0f: Adds `Legend` component
+
+### Patch Changes
+
+- Updated dependencies [206278b0f]
+- Updated dependencies [206278b0f]
+- Updated dependencies [206278b0f]
+  - @lg-charts/colors@0.2.2
+  - @leafygreen-ui/checkbox@15.2.0
+  - @lg-charts/series-provider@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes