+++ packages/drawer/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @leafygreen-ui/drawer
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [30b13adec]
+- Updated dependencies [30b13adec]
+- Updated dependencies [78a36d6bb]
+  - @leafygreen-ui/hooks@8.4.0
+  - @leafygreen-ui/lib@14.2.0
+  - @leafygreen-ui/emotion@4.1.0
+  - @leafygreen-ui/leafygreen-provider@4.0.6
+  - @leafygreen-ui/tabs@14.2.2
+  - @leafygreen-ui/button@23.1.4
+  - @leafygreen-ui/icon@13.2.1
+  - @leafygreen-ui/icon-button@16.0.8
+  - @leafygreen-ui/palette@4.1.4
+  - @leafygreen-ui/polymorphic@2.0.8
+  - @leafygreen-ui/tokens@2.12.1
+  - @leafygreen-ui/typography@20.1.6
+
 ## 2.0.0
 
 ### Major Changes