+++ packages/side-nav/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @leafygreen-ui/side-nav
 
+## 16.0.10
+
+### Patch Changes
+
+- Updated dependencies [30b13adec]
+- Updated dependencies [30b13adec]
+- Updated dependencies [78a36d6bb]
+  - @leafygreen-ui/hooks@8.4.0
+  - @leafygreen-ui/lib@14.2.0
+  - @leafygreen-ui/emotion@4.1.0
+  - @leafygreen-ui/a11y@2.0.5
+  - @leafygreen-ui/leafygreen-provider@4.0.6
+  - @leafygreen-ui/portal@6.0.5
+  - @leafygreen-ui/tooltip@13.0.9
+  - @leafygreen-ui/icon@13.2.1
+  - @leafygreen-ui/palette@4.1.4
+  - @leafygreen-ui/polymorphic@2.0.8
+  - @leafygreen-ui/tokens@2.12.1
+  - @leafygreen-ui/typography@20.1.6
+
 ## 16.0.9
 
 ### Patch Changes