+++ packages/split-button/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @leafygreen-ui/split-button
 
+## 4.1.10
+
+### Patch Changes
+
+- Updated dependencies [30b13adec]
+- Updated dependencies [30b13adec]
+- Updated dependencies [78a36d6bb]
+- Updated dependencies [30b13adec]
+  - @leafygreen-ui/hooks@8.4.0
+  - @leafygreen-ui/lib@14.2.0
+  - @leafygreen-ui/emotion@4.1.0
+  - @leafygreen-ui/menu@29.0.0
+  - @leafygreen-ui/leafygreen-provider@4.0.6
+  - @leafygreen-ui/popover@13.0.8
+  - @leafygreen-ui/button@23.1.4
+  - @leafygreen-ui/icon@13.2.1
+  - @leafygreen-ui/palette@4.1.4
+  - @leafygreen-ui/polymorphic@2.0.8
+  - @leafygreen-ui/tokens@2.12.1
+
 ## 4.1.9
 
 ### Patch Changes