+++ packages/table/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @leafygreen-ui/table
 
+## 13.1.5
+
+### Patch Changes
+
+- Updated dependencies [30b13adec]
+- Updated dependencies [30b13adec]
+- Updated dependencies [78a36d6bb]
+  - @leafygreen-ui/hooks@8.4.0
+  - @leafygreen-ui/lib@14.2.0
+  - @leafygreen-ui/emotion@4.1.0
+  - @leafygreen-ui/checkbox@15.2.1
+  - @leafygreen-ui/leafygreen-provider@4.0.6
+  - @leafygreen-ui/icon@13.2.1
+  - @leafygreen-ui/icon-button@16.0.8
+  - @leafygreen-ui/palette@4.1.4
+  - @leafygreen-ui/polymorphic@2.0.8
+  - @leafygreen-ui/tokens@2.12.1
+  - @leafygreen-ui/typography@20.1.6
+
+## 13.1.4
+
+### Patch Changes
+
+- Updated dependencies [206278b0f]
+  - @leafygreen-ui/checkbox@15.2.0
+
 ## 13.1.3
 
 ### Patch Changes